py-backend/requirements.txt
fastapi
uvicorn
python-chess  # To handle chess rules and validation
